+++ packages/test-apps/react-vite/src/components/ColorfulBox/ColorfulBox.module.css
